requirements.txt
Django~=5.0.7
djangorestframework~=3.15.2
pytz~=2023.3
tzdata~=2022.1
django-cors-headers~=3.10.1